requirements/requirements.txt
pydantic
requests
python-dotenv
langchain
langchain-community 
langchain-core